+++ pychron/experiment/automated_run/measurement_fits_selector.py
@@ -69,7 +69,7 @@ class MeasurementFitsSelector(FilterFitSelector):
         return ys
 
     def save(self):
-        sfs, bfs = partition(self.fits, lambda x: x.is_baseline)
+        bfs, sfs = partition(self.fits, lambda x: x.is_baseline)
         yd = {'signal': self._dump(sfs),
               'baseline': self._dump(bfs)}