+++ pychron/experiment/tasks/experiment_editor.py
@@ -15,7 +15,7 @@
 #===============================================================================
 
 #============= enthought library imports =======================
-from traits.api import Instance, Unicode, Property, DelegatesTo
+from traits.api import Instance, Unicode, Property, DelegatesTo, Color
 from traitsui.api import View, UItem
 
 #============= standard library imports ========================
@@ -49,6 +49,12 @@ class ExperimentEditor(BaseTraitsEditor):
     executed = DelegatesTo('queue')
     tabular_adapter_klass = AutomatedRunSpecAdapter
     executed_tabular_adapter_klass = ExecutedAutomatedRunSpecAdapter
+    bgcolor = Color
+
+    def set_bgcolor(self, c):
+        self.bgcolor=c
+        self.tabular_adapter_klass.odd_bg_color=c
+        self.executed_tabular_adapter_klass.odd_bg_color=c
 
     def new_queue(self, txt=None, **kw):
         queue = self.queue_factory(**kw)
@@ -78,6 +84,7 @@ class ExperimentEditor(BaseTraitsEditor):
                          editor=myTabularEditor(adapter=self.tabular_adapter_klass(),
                                                 operations=['delete',
                                                             'move', 'edit'],
+                                                bgcolor=self.bgcolor,
                                                 editable=True,
                                                 show_row_titles=True,
                                                 dclicked='dclicked',
@@ -91,6 +98,7 @@ class ExperimentEditor(BaseTraitsEditor):
 
         executed_grp = UItem('executed_runs',
                              editor=myTabularEditor(adapter=self.executed_tabular_adapter_klass(),
+                                                    bgcolor=self.bgcolor,
                                                     editable=False,
                                                     auto_update=True,
                                                     selectable=True,